+++ src/config.ts
@@ -1,3 +1,4 @@
+import {homedir} from 'os';
 import * as path from 'path';
 import * as vscode from 'vscode';
 
@@ -34,6 +35,9 @@ function substitute<T>(val: T): T {
 // Subset of substitution variables that are most likely to be useful.
 // https://code.visualstudio.com/docs/editor/variables-reference
 function replacement(name: string): string|undefined {
+  if (name === 'userHome') {
+    return homedir();
+  }
   if (name === 'workspaceRoot' || name === 'workspaceFolder' ||
       name === 'cwd') {
     if (vscode.workspace.rootPath !== undefined)